+++ packages/react/src/DevMode-production.spec.tsx
@@ -1,7 +1,0 @@
-import { ProductionDevMode } from './contexts/SuspensiveDevModeContext'
-
-describe('<DevMode/> in production mode', () => {
-  it('should show nothing if without SuspensiveProvider', () => {
-    expect(ProductionDevMode()).toBeNull()
-  })
-})