+++ packages/react/src/contexts/index.ts
@@ -1,2 +1,1 @@
 export { DelayDefaultPropsContext, SuspenseDefaultPropsContext } from './DefaultPropsContexts'
-export { SuspensiveDevMode, DevModeContext, syncDevMode } from './SuspensiveDevModeContext'