+++ packages/react-query-4/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @suspensive/react-query-4
 
+## 2.9.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.9.3
+  - @suspensive/react@2.9.3
+
+## 2.9.2
+
+### Patch Changes
+
+- Updated dependencies [[`89a6528`](https://github.com/toss/suspensive/commit/89a652866699f09853f9baf74a973767f2350661)]:
+  - @suspensive/react@2.9.2
+  - @suspensive/utils@2.9.2
+
+## 2.9.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @suspensive/utils@2.9.1
+  - @suspensive/react@2.9.1
+
 ## 2.9.0
 
 ### Patch Changes